Add files via upload
The hospital management system offers administrators, medical personnel, and patients a seamless experience by effectively managing patient records, staff information, and hospital-related activities. It enables the keeping, updating, and searching of staff and patient. This system makes sure that data is
updated, accessed with ease, and kept in a methodical manner. It facilitates adding new employees and patients, editing personal information, accessing all records, and creating reports. It can also load and save records to and from a file, guaranteeing trustworthy data handling. The system improves hospital
information management's dependability and efficiency, which helps to make the healthcare process more efficient and well-organized. Finally, the patient has the option to submit a review and rate the hospital.